backend/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Flask==3.0.0
flask-cors==4.0.0
vonage==3.1.1